+++ packages/expo-widgets/src/Widgets.types.ts
@@ -307,6 +307,7 @@ export declare class NativeLiveActivityFactory extends SharedObject {
 }
 
 export declare class NativeLiveActivity extends SharedObject<LiveActivityEvents> {
+  getId(): string;
   update(props?: string, staleDate?: number): Promise<void>;
   end(
     dismissalPolicy?: string,